+++ resources/views/articles/show.blade.php
@@ -19,8 +19,10 @@
         <div class="card-body">
             <h5 class="card-title">{{$comment->title}}  ({{$comment->created_at}})</h5>
             <p class="card-text">{{$comment->text}}</p>
-            <a href="/comment/{{$comment->id}}/edit" class="btn btn-secondary">Редактировать</a>
-            <a href="/comment/{{$comment->id}}/delete" class="btn btn-secondary">Удалить</a>
+            @can('update-comment', $comment)
+                <a href="/comment/{{$comment->id}}/edit" class="btn btn-secondary">Редактировать</a>
+                <a href="/comment/{{$comment->id}}/delete" class="btn btn-secondary">Удалить</a>
+            @endcan
         </div>
     @endforeach
     </div>
